Rename the spatial hash files
