+++ simulant/partitioners/impl/spatial_hash.h
