+++ simulant/partitioners/spatial_hash.h
@@ -2,7 +2,7 @@
 
 
 #include "../partitioner.h"
-#include "./impl/hgsh.h"
+#include "./impl/spatial_hash.h"
 #include "../stage.h"
 #include "../generic/threading/shared_mutex.h"
